+++ src/Lykke.RabbitMqBroker/Subscriber/RabbitMqListenerOptions.cs
@@ -9,37 +9,38 @@ namespace Lykke.RabbitMqBroker.Subscriber
         /// Instructs the listener to request shared connection instead of creating a new one.
         /// </summary>
         public bool ShareConnection { get; set; } = true;
-        
+
         /// <summary>
         /// Creates subscriber(-s) with JSON serialization format by default.
         /// </summary>
         public SerializationFormat SerializationFormat { get; set; } = SerializationFormat.Json;
-        
+
         /// <summary>
         /// Creates subscriber(-s) with "no loss" subscription template by default.
         /// </summary>
         public SubscriptionTemplate SubscriptionTemplate { get; set; } = SubscriptionTemplate.NoLoss;
-        
+
+
         /// <summary>
         /// The number of low-level subscribers to create.
         /// </summary>
-        public byte ConsumerCount { get; set; } = 1;
-        
+        public ConsumerCount ConsumerCount { get; set; } = ConsumerCount.Default;
+
         public static class Json
         {
             public static RabbitMqListenerOptions<T> Default => NoLoss;
-            
+
             public static RabbitMqListenerOptions<T> NoLoss => new RabbitMqListenerOptions<T>
             {
-                ConsumerCount = 1,
+                ConsumerCount = ConsumerCount.Default,
                 SerializationFormat = SerializationFormat.Json,
                 ShareConnection = true,
                 SubscriptionTemplate = SubscriptionTemplate.NoLoss
             };
-            
+
             public static RabbitMqListenerOptions<T> LossAcceptable => new RabbitMqListenerOptions<T>
             {
-                ConsumerCount = 1,
+                ConsumerCount = ConsumerCount.Default,
                 SerializationFormat = SerializationFormat.Json,
                 ShareConnection = true,
                 SubscriptionTemplate = SubscriptionTemplate.LossAcceptable
@@ -49,10 +50,10 @@ namespace Lykke.RabbitMqBroker.Subscriber
         public static class MessagePack
         {
             public static RabbitMqListenerOptions<T> Default => NoLoss;
-            
+
             public static RabbitMqListenerOptions<T> NoLoss => new RabbitMqListenerOptions<T>
             {
-                ConsumerCount = 1,
+                ConsumerCount = ConsumerCount.Default,
                 SerializationFormat = SerializationFormat.Messagepack,
                 ShareConnection = true,
                 SubscriptionTemplate = SubscriptionTemplate.NoLoss
@@ -60,7 +61,7 @@ namespace Lykke.RabbitMqBroker.Subscriber
 
             public static RabbitMqListenerOptions<T> LossAcceptable => new RabbitMqListenerOptions<T>
             {
-                ConsumerCount = 1,
+                ConsumerCount = ConsumerCount.Default,
                 SerializationFormat = SerializationFormat.Messagepack,
                 ShareConnection = true,
                 SubscriptionTemplate = SubscriptionTemplate.LossAcceptable